changed main file in inheritance branch. new changes in Student and Instructor code

+++ src/main/java/org/example/service/Calculator.java
@@ -17,7 +17,7 @@ public class Calculator {
         answer = a - b;
     }
 
-    public void division(double a, double b){
+    public void division(int a, int b){
         answer = a / b;
     }
